added support for demo
+ moved utility functions out of demo
+ addded support for voltage imaging in demo
